adicionado validação de saldo de produtos e serviços, botão de reset de quantidade e ajustes nos selects

+++ sistema/ordemcompraProduto.php
@@ -21,34 +21,58 @@ if(isset($_POST['inputOrdemCompraId'])){
 
 //Se está alterando
 if(isset($_POST['inputIdOrdemCompra'])){
-	
-	$sql = "DELETE FROM OrdemCompraXProduto
-			WHERE OCXPrOrdemCompra = :iOrdemCompra AND OCXPrUnidade = :iUnidade";
-	$result = $conn->prepare($sql);
-	
-	$result->execute(array(
-					':iOrdemCompra' => $iOrdemCompra,
-					':iUnidade' => $_SESSION['UnidadeId']
-					));		
-	
-	for ($i = 1; $i <= $_POST['totalRegistros']; $i++) {
-	
-		$sql = "INSERT INTO OrdemCompraXProduto (OCXPrOrdemCompra, OCXPrProduto, OCXPrQuantidade, OCXPrValorUnitario, OCXPrUsuarioAtualizador, OCXPrUnidade)
-				VALUES (:iOrdemCompra, :iProduto, :iQuantidade, :fValorUnitario, :iUsuarioAtualizador, :iUnidade)";
+	$valid = true;
+
+	for ($i = 1; $i <= $_POST['totalRegistros']; $i++){
+		$sqlSaldo = "SELECT dbo.fnSaldoOrdemCompra($_SESSION[UnidadeId], '$iOrdemCompra', ".$_POST['inputIdProduto'.$i].", 'P') as Saldo";
+		$resultSaldo = $conn->query($sqlSaldo);
+		$saldo = $resultSaldo->fetch(PDO::FETCH_ASSOC);
+
+		if(intval($_POST['inputQuantidade'.$i]) > intval($saldo['Saldo'])){
+			$valid = false;
+		}
+	}
+	if($valid){
+		$sql = "DELETE FROM OrdemCompraXProduto
+				WHERE OCXPrOrdemCompra = :iOrdemCompra AND OCXPrUnidade = :iUnidade";
 		$result = $conn->prepare($sql);
 		
 		$result->execute(array(
 						':iOrdemCompra' => $iOrdemCompra,
-						':iProduto' => $_POST['inputIdProduto'.$i],
-						':iQuantidade' => $_POST['inputQuantidade'.$i] == '' ? null : $_POST['inputQuantidade'.$i],
-						':fValorUnitario' => $_POST['inputValorUnitario'.$i] == '' ? null : gravaValor($_POST['inputValorUnitario'.$i]),
-						':iUsuarioAtualizador' => $_SESSION['UsuarId'],
 						':iUnidade' => $_SESSION['UnidadeId']
 						));
 		
-		$_SESSION['msg']['titulo'] = "Sucesso";
-		$_SESSION['msg']['mensagem'] = "Ordem de Compra alterado!!!";
-		$_SESSION['msg']['tipo'] = "success";
+		for ($i = 1; $i <= $_POST['totalRegistros']; $i++) {
+			// verificar se o saldo ainda é valido antes de inserir no banco os valores
+			$sqlSaldo = "SELECT dbo.fnSaldoOrdemCompra($_SESSION[UnidadeId], '$iOrdemCompra', ".$_POST['inputIdProduto'.$i].", 'P') as Saldo";
+			$resultSaldo = $conn->query($sqlSaldo);
+			$saldo = $resultSaldo->fetch(PDO::FETCH_ASSOC);
+
+			$saldo = $saldo['Saldo'];
+			$quantidade = $_POST['inputQuantidade'.$i];
+
+			
+				$sql = "INSERT INTO OrdemCompraXProduto (OCXPrOrdemCompra, OCXPrProduto, OCXPrQuantidade, OCXPrValorUnitario, OCXPrUsuarioAtualizador, OCXPrUnidade)
+						VALUES (:iOrdemCompra, :iProduto, :iQuantidade, :fValorUnitario, :iUsuarioAtualizador, :iUnidade)";
+				$result = $conn->prepare($sql);
+				
+				$result->execute(array(
+								':iOrdemCompra' => $iOrdemCompra,
+								':iProduto' => $_POST['inputIdProduto'.$i],
+								':iQuantidade' => $_POST['inputQuantidade'.$i] == '' ? null : $_POST['inputQuantidade'.$i],
+								':fValorUnitario' => $_POST['inputValorUnitario'.$i] == '' ? null : gravaValor($_POST['inputValorUnitario'.$i]),
+								':iUsuarioAtualizador' => $_SESSION['UsuarId'],
+								':iUnidade' => $_SESSION['UnidadeId']
+								));
+				
+				$_SESSION['msg']['titulo'] = "Sucesso";
+				$_SESSION['msg']['mensagem'] = "Ordem de Compra alterado!!!";
+				$_SESSION['msg']['tipo'] = "success";
+		}
+	}else{
+		$_SESSION['msg']['titulo'] = "Erro";
+		$_SESSION['msg']['mensagem'] = "O saldo de um dos produtos não está mais disponível!!!";
+		$_SESSION['msg']['tipo'] = "error";
 	}
 }	
 
@@ -196,6 +220,10 @@ try{
 		function ResetProduto(){
 			$('#cmbProduto').empty().append('<option>Sem produto</option>');
 		}
+
+		function reset(id){
+			confirmaReset(document.formOrdemCompraProduto, "Tem certeza que deseja resetar essa quantidade?", "ordemcompraProduto.php", "inputQuantidade"+id);
+		}
 		
 		function calculaValorTotal(id){
 			
@@ -216,8 +244,7 @@ try{
 			$('#inputTotalGeral').val(TotalGeral);			
 		}
 		function validaQuantInputModal(quantMax, obj) {
-			console.log(obj.id)
-				$('#'+obj.id).on('keyup', function() {
+			$('#'+obj.id).on('keyup', function() {
 					if (parseInt($('#'+obj.id).val()) > parseInt(quantMax)) {
 						$('#'+obj.id).val(quantMax)
 					}
@@ -307,19 +334,21 @@ try{
 																FROM Produto
 																JOIN Situacao on SituaId = ProduStatus
 																JOIN FluxoOperacionalXProduto on FOXPrProduto = ProduId and FOXPrFluxoOperacional = '$iOrdemCompraFlOpe'
-																WHERE ProduUnidade = $_SESSION[UnidadeId] and SituaChave = 'ATIVO'";
-														
+																WHERE ProduUnidade = $_SESSION[UnidadeId] and SituaChave = 'ATIVO' and
+																ProduCategoria = '$iCategoria'";
+														if (isset($row['OrComSubCategoria']) and $row['OrComSubCategoria'] != '' and $row['OrComSubCategoria'] != null){
+															$sql .= " and ProduSubCategoria = ".$row['OrComSubCategoria'];
+														}
 														$sql .= " ORDER BY ProduNome ASC";
 														$result = $conn->query($sql);
-														$rowProduto = $result->fetchAll(PDO::FETCH_ASSOC);														
-														
+														$rowProduto = $result->fetchAll(PDO::FETCH_ASSOC);
 														foreach ($rowProduto as $item){
 															
 															if (in_array($item['ProduId'], $aProdutos) or $countProdutoUtilizado == 0) {
 																$seleciona = "selected";
 															} else {
 																$seleciona = "";
-															}													
+															}
 															
 															print('<option value="'.$item['ProduId'].'" '.$seleciona.'>'.$item['ProduNome'].'</option>');
 														}
@@ -361,6 +390,9 @@ try{
 													JOIN UnidadeMedida on UnMedId = ProduUnidadeMedida
 													JOIN FluxoOperacionalXProduto on FOXPrProduto = ProduId and FOXPrFluxoOperacional = '$iOrdemCompraFlOpe'
 													WHERE ProduUnidade = ".$_SESSION['UnidadeId']." and SituaChave = 'ATIVO'";
+										if (isset($row['OrComSubCategoria']) and $row['OrComSubCategoria'] != '' and $row['OrComSubCategoria'] != null){
+											$sql .= " and ProduSubCategoria = ".$row['OrComSubCategoria'];
+										}
 										$sql = $sql." ORDER BY ProduNome";
 										$result = $conn->query($sql);
 										$rowProdutos = $result->fetchAll(PDO::FETCH_ASSOC);
@@ -373,6 +405,9 @@ try{
 															JOIN UnidadeMedida on UnMedId = ProduUnidadeMedida
 															JOIN FluxoOperacionalXProduto on FOXPrProduto = ProduId and FOXPrFluxoOperacional = '$iOrdemCompraFlOpe'
 															WHERE ProduUnidade = ".$_SESSION['UnidadeId']." and SituaChave = 'ATIVO'";
+											if (isset($row['OrComSubCategoria']) and $row['OrComSubCategoria'] != '' and $row['OrComSubCategoria'] != null){
+												$sql .= " and ProduSubCategoria = ".$row['OrComSubCategoria'];
+											}
 											$sql = $sql." ORDER BY ProduNome";
 											$result = $conn->query($sql);
 											$rowProdutos = $result->fetchAll(PDO::FETCH_ASSOC);
@@ -381,12 +416,12 @@ try{
 										
 										print('
 										<div class="row" style="margin-bottom: -20px;">
-											<div class="col-lg-6">
+											<div class="col-lg-5">
 												<div class="row">
-													<div class="col-lg-2">
+													<div class="col-lg-2" style="max-width:60px">
 														<label for="inputCodigo"><strong>Item</strong></label>
 													</div>
-													<div class="col-lg-10">
+													<div class="col-lg-10" style="width:100%">
 														<label for="inputProduto"><strong>Produto</strong></label>
 													</div>
 												</div>
@@ -415,7 +450,12 @@ try{
 												<div class="form-group">
 													<label for="inputValorTotal"><strong>Valor Total</strong></label>
 												</div>
-											</div>											
+											</div>
+											<div class="col-sm-1">
+												<div class="col-sm-12" style="text-align:center;">
+													<label for=""><strong>Resetar</strong></label>
+												</div>
+											</div>
 										</div>');
 										
 										print('<div id="tabelaProdutos">');
@@ -424,22 +464,22 @@ try{
 										
 										foreach ($rowProdutos as $item){
 											$cont++;
-											$saldo = isset($item['SaldoOrdemCompra']) ? $item['SaldoOrdemCompra'] : '';
-											$iQuantidade = isset($item['OCXPrQuantidade']) ? $item['OCXPrQuantidade'] : '';
-											$fValorUnitario = isset($item['FOXPrValorUnitario']) ? mostraValor($item['FOXPrValorUnitario']) : mostraValor($item['OCXPrValorUnitario']);											
-											$fValorTotal = (isset($item['OCXPrQuantidade']) and isset($item['OCXPrValorUnitario'])) ? mostraValor($item['OCXPrQuantidade'] * $item['OCXPrValorUnitario']) : '';
+											$saldo = isset($item['SaldoOrdemCompra']) ? $item['SaldoOrdemCompra'] : 0;
+											$iQuantidade = isset($item['OCXPrQuantidade']) ? $item['OCXPrQuantidade'] : 0;
+											$fValorUnitario = isset($item['FOXPrValorUnitario']) ? mostraValor($item['FOXPrValorUnitario']) : 0;											
+											$fValorTotal = mostraValor(intval($iQuantidade)*gravaValor($fValorUnitario));
 											
-											$fTotalGeral += (isset($item['OCXPrQuantidade']) and isset($item['OCXPrValorUnitario'])) ? $item['OCXPrQuantidade'] * $item['OCXPrValorUnitario'] : 0;
+											$fTotalGeral += gravaValor($fValorTotal);
 											
 											print('
 											<div class="row" style="margin-top: 8px;">
-												<div class="col-lg-6">
+												<div class="col-lg-5">
 													<div class="row">
-														<div class="col-lg-2">
+														<div class="col-lg-2" style="max-width:60px">
 															<input type="text" id="inputItem'.$cont.'" name="inputItem'.$cont.'" class="form-control-border-off" value="'.$cont.'" readOnly>
 															<input type="hidden" id="inputIdProduto'.$cont.'" name="inputIdProduto'.$cont.'" value="'.$item['ProduId'].'" class="idProduto">
 														</div>
-														<div class="col-lg-10">
+														<div class="col-lg-10" style="width:100%">
 															<input type="text" id="inputProduto'.$cont.'" name="inputProduto'.$cont.'" class="form-control-border-off" data-popup="tooltip" title="'.$item['ProduDetalhamento'].'" value="'.$item['ProduNome'].'" readOnly>
 														</div>
 													</div>
@@ -451,16 +491,18 @@ try{
 													<input type="text" id="inputSaldo'.$cont.'" readOnly name="Saldo'.$cont.'" class="form-control-border-off text-right" value="'.$saldo.'">
 												</div>
 												<div class="col-lg-1">
-													<input type="text" class="form-control-border Quantidade text-right" id="inputQuantidade'.$cont.'" '.($saldo > 0?'':'readOnly').' name="inputQuantidade'.$cont.'" onkeypress="validaQuantInputModal('.$saldo.',this)" onChange="calculaValorTotal('.$cont.')" onkeypress="return onlynumber();" value="'.$iQuantidade.'">
+													<input type="text" class="form-control-border Quantidade text-right" id="inputQuantidade'.$cont.'" '.($saldo > 0?'':'readOnly').' name="inputQuantidade'.$cont.'" onChange="calculaValorTotal('.$cont.')" onkeypress="return onlynumber(), validaQuantInputModal('.$saldo.',this)" value="'.$iQuantidade.'">
 												</div>		
 												<div class="col-lg-1">
 													<input readOnly type="text" id="inputValorUnitario'.$cont.'" name="inputValorUnitario'.$cont.'" class="form-control-border-off ValorUnitario text-right" onChange="calculaValorTotal('.$cont.')" onKeyUp="moeda(this)" maxLength="12" value="'.$fValorUnitario.'">
 												</div>	
 												<div class="col-lg-2">
 													<input type="text" id="inputValorTotal'.$cont.'" name="inputValorTotal'.$cont.'" class="form-control-border-off text-right" value="'.$fValorTotal.'" readOnly>
-												</div>											
-											</div>');											
-											
+												</div>
+												<div class="col-sm-1 btn" style="text-align:center;" onClick="reset('.$cont.')">
+													<i class="icon-reset" title="Resetar"></i>
+												</div>
+											</div>');
 										}
 										
 										print('
@@ -484,7 +526,7 @@ try{
 												<div class="col-lg-1">
 													
 												</div>	
-												<div class="col-lg-2" style="padding-top: 5px; text-align: right;">
+												<div class="col-lg-1" style="padding-top: 5px; text-align: right;">
 													<h5><b>Total:</b></h5>
 												</div>	
 												<div class="col-lg-2">
